+++ eggs/Naaya/eggs/naaya.groupware/CHANGELOG.rst
@@ -1,5 +1,6 @@
 1.2.6 (unreleased)
 -------------------
+* explanatory text for ig membership request [dumitval]
 
 1.2.5 (2012-05-14)
 -------------------